[FIX] ExtensionPoint: fixed non-working ExtensionPoints in Fragments
Corrected the require statement for Fragment:
Fragments are located in sap.ui.core and not in sap.ui.core.mvc and
therfore the require is adopted. Right now the extension points for
Fragments do not work because of the wrong import.

Enhanced the QUnit tests to check the ExtensionPoint in Fragments to
work properly to avoid such kind of issues in future.

Change-Id: I5bbd2dc5e1490a416fab1cc9071d8adabd3a14cc
BCP: 012002523100005473112015