[INTERNAL] jquery.sap.global: Allow function as assert message
This change allows givin funtions to jQuery.sap.assert instead of
strings, this can be used if the content of the message is calculated
on the fly. If for example an error occurs when pasring an element in
a DOM tree and with the failed assertion we want to add the XPath 
describing the element to the message, this path would be calculated
even when the assertion does not fail.

Change-Id: I4020738d101c908b1c0792d0ffba67b1a7df9cfc