[FIX] ux3.Shell: Set a static z-index for NotificationBar
Scenario:
If there is a NotiBar attached to a Shell and the Shell-Content contains
a ThingInspector and within the ThingInspector there is a ToolPopup that
opens a Datepicker.

If all popups (ThingInspector, ToolPopup and Datepicker) are open and
the focus returns into the Datepicker's input field the z-index of all
popups will be increased. If the NotiBar has been opened already it will
be covered by all Popups due to the increased z-index.

So the NotiBar within the Shell will be set a z-index of 100.000 and
this should be enough until there is a smart solution for the popup.

BCP: 1570523041
CR-Id: 012006153200003222092015
Change-Id: Ib1c77e18b5d293bdc79b7cb4e75ff5aea5e26d74
(cherry picked from commit 409544fe0ee240e4d9d217ce1856f2a428947ff9)