[FIX] Destroy templates automatically in more cases.
When a template is neither marked as "shareable" nor as "not shareable",
mark it as a candidate for automatic deletion. When a new element with
the same ID is about to being created, destroy the old one first.
If an owning component is known and is about to being destroyed, destroy
any such templates as well. 

Still, templateShareable:true prevents deletion of templates and
templateShareable:false enforces it.

Change-Id: Iacbc82854d5b5546c1d23e21f9f910b059bc2a00