[INTERNAL] DiscoveryServlet: added a cast check for the file field
In the listClasspathResources function we list the resources inside a
folder of a JAR entry or a folder in the file system. If running the
DiscoveryServlet in JBoss the value of the file field is a VirtualFile
which cannot be casted to a File. To avoid the exception we do a cast
check before now. 

This issue has been reported with #518. This fix is only polishing the
exception to avoid it but a final solution would be to extend the
DiscoveryServlet to allow to do a custom implementation of the resource
listing per container.

Change-Id: I9f69aa15eb8b3687c8c73b798562de4d991ab917