[INTERNAL] TestSuite: support to start a Jetty for testing purposes
The profile "Jetty" in the project "testsuite" can be used to run an
embedded Jetty with the UI5 testsuite. To start the Jetty you can run
the following Maven command: 

	mvn clean integration-test -P jetty

The Jetty will be started on port 8080 and block the console as long as
the Jetty is running and can be stopped with CTRL+C.

BUT: before this will work you need to install the UI5 libraries into
your local Maven repository once with: mvn clean install within the
reactor of the openui5 project in the root folder of this Git
repository.

In order to execute the build and run the Jetty afterwards you can do so
by executing the following Maven command in the root folder of this Git
repository: 

	mvn clean integration-test -P jetty,testsuite.build

This will build all UI5 libraries and the Java artifacts and finally the
build results will be used for the deployment into the Jetty server.
Again you can stop the Jetty with CTRL-C once you need to rebuild or
retest.

INFO: since this server is using the build artifacts changes in the
source code will not directly get reflected in the server without
rebuilding the UI5 library (mvn clean install) which has been changed or
execute the build on the full project.

Change-Id: Iad8d0df3e22fef16f84346ff62e47cd2264a443c