[FIX] Popup: Resize Directly After Open
If the opener was focused right after the popup has opened but the
opener exceeds the current window width the browser will
scroll/reposition accordingly. When this popup registers the
followOf-Handler the check if the opener moved will result in that the
opener moved due to the focus and scrolling of the browser. So it is
necessary to resize/reposition the popup right after the focus.

BCP: 012002523100005790342015
CR-Id: 012006153200003528512015
Change-Id: If40c167e27400faafa9d1358519a86587d508b76