[FEATURE] v2.ODataListBinding.js - OperationMode.Auto
For CountMode.Request/Both. the OperationMode.Auto first requests a
$count and then compares the count against the given threshold. If the
threshold is lower -> all data will be loaded and the binding behaves
like the OperationMode.Client, else it behaves like the
OperationMode.Server.

For CountMode.Inline, the first request tries to load as many entries as
the given threshold. The decision to operate locally or serverside is
analogue to the CountMode.Request.

Change-Id: Ibc3f51a2a7572f25d649428d41fae76c7213dac1