[FIX] sap.ui.test.Opa5 - IFrame navigation
When a user was navigating in the following flow
PageA -> PageB -> Browser Back -> PageB
the last hashchange to PageB was ignored.

This commit should also fix the app to app navigation that apparently noone tests with OPA:
Hashchange events are now caught one level deeper -> hasher instead of HashChanger this includes the whole window hash when running in a launchpad.

BCP: 1570630363
Change-Id: Ibf68e87f226f79017a9f360da48de03e283d69f6