[FIX] sap.m.Popover: Now registers itself as a child popup correctly
If Popover is opened from another Popover and the child Popover is with
placement type different from Left, Right, Top or Bottom, the child
Popover is not registered as child of the parent at first opening.

This is due to the reason that in openBy method oPopup.setPosition method is
called only when placement type is Left, Right, Top or Bottom.

However here we should add the correct 'of' element, because if we do not
set it at this point the dafault 'of' (document element) is used in methods that
check if this Popover has parent Popover and the result is always false.

The right 'of' is set afterwards when calcPlacement method is called, this
is the reason why on the second opening of the child Popover, it
sucessfully register itself as a child.

Change-Id: I0c5b7bc1a5b27db292d27ca54ed9ae7dd2a4a28b
BCP: 1570771012
CR-Id: 012006153200003800072015