[INTERNAL] sap.m.Title: Added proper visualization of title in info toolbar
This is a fix in response to BCP ticket
https://support.wdf.sap.corp/sap/support/message/1570761773. The problem
is:

When you use a sap.m.Title to display text in an infoToolbar, the
styling is not according to specification:
https://wiki.wdf.sap.corp/wiki/display/visualcore/ContextualFilter

No such problem if you use sap.m.Label or sap.m.Text instead of
sap.m.Title.

Assuming this is a quite possible use-case, I added support for styling
of title in info-toolbar context.

Change-Id: I82a060edf1f16045f8b890b623271b887416c7c9