[FIX] ManagedObject: support binding expressions with an empty path
An empty String is a valid binding path which needs to be considered
when extracting the binding information in the
ManagedObject.prototype.applySettings.

Binding Expressions in e.g. XMLViews should also consider an empty path
as a kind of a dummy binding to allow to define a template for the
binding as content of an aggregation. Especially if such kind of
bindings should be established dynamically later and the template should
be defined in a declarative way this is one option.

Change-Id: I9c75da900c552f38ec02f4f09e6b91e46ac1f15b
BCP: 1570774108
CR-Id: 012006153200004400912015