[INTERNAL] Add library / dist version to library.css URI
To enable custom theme compilation on demand, the library / dist
version needs to be passed to the server.

This change adds a new "sap-ui-versionedLibCss" configuration.
The core bootstrap loads the "sap-ui-version.json" if the
configuration is set to "true". This request will be sync/async
according to the "sap-ui-preload" configuration option
(sync by default).

The following theme resource will then be loaded with
the library and dist version as a URI parameter:
- library(-RTL).css (or any other variation)
- library-parameters.json

URLs within the CSS or Parameters are not modified.

Change-Id: I6609cf379a3d255bced6440fd7a88dbd4dd05ba6