[FIX] Popup: Using FocusInfo Fixed
There are two different scenarios using the FocusInfo:
- Keep the value inside an input field if the renderer re-renders the
input
- The Popup focuses the previous focused control/element and uses the
FocusInfo mechanism as well.

If the FocusInfo is used by the Popup the Popup marks the FocusInfo and
controls could react on this - the ComboBox for example. So the ComboBox
is able to keep the current value inside the input field that might be
changed by an application inside the Popup.

CR-Id: 012003146900002992022015
BCP: 0020283765 0000007099 2015
BCP: 0020283766 0000045097 2015
Change-Id: I2d4f57ad2de9ea3399940d382714b037da5eb8e5
(cherry picked from commit 8e7bac69dbad1a513bfa73ef589323f2a584bc7e)