[INTERNAL] sap.ui.core.delegate.ScrollEnablement - apply scroll state only if needed to improve rendering performance
Rendering performance of controls relying on ScrollEnablement will be improved as
any reading from layout immediately after the onAfterRendering of ScrollEnablement
will not trigger layout recalculation.
This will have a positive impact to the initial rendering of controls that are
based on ScrollEnablement and, when showing for the first time, do not have
any scroll state.

Change-Id: I1f1e93267d5c13d54dc8685fb5ec27bf873e37ee