[FIX] sap.m.Select: cannot filter binding in change event handler
If the items of the sap.m.Select control are filter in the change event
and the selected item does not changed, the change event should not be
fired. Now this issue is fixed, by synchronizing the selection after the
items are filtered.

Change-Id: I7467b07681d6d1e69969c53a8f7dcb335d9b0ac4
BCP: 1570825099
(cherry picked from commit 2f3ca10b079bb0a620aead5590f41c65a81366f9)
CR-Id: 012003146900003128662015