[INTERNAL][FIX] Corrected font-size of Text, Label and Title in Toolbar
Fixed the problem (BCP 1570801016) with normal labels and text sizes
getting over styled in toolbars with the larger title sizes.

According to spec, the title sizes should only be used on specified
titles and headers, so normal labels and texts should not be enlarged

A problem could occur that someone places a label beside an input and
the label gets larger than the default label and text size.

Change-Id: I67b7d4f11608c09c86b15ca8be6bdef9a18b0f59
BCP: 1570801016