[FIX] XMLView: ensure that content is preserved during rendering
For the preservation of pure HTML content of an XMLView, it is 
important that the XMLView finds the preserved DOM during rendering. 
For the standard, async rendering triggered via invalidate(), this is 
ensured by the default RenderManager. But for a custom RenderManager
created with Core.createRenderManager(), the DOM preservation is done
only after the XMLView rendering when the render buffer is flushed to
the DOM. As a result, the content of the XMLView was rendered a second
time and the old content was kept in the preserve area (duplicate DOM
IDs!).

With this change, the XMLView ensures that the DOM is preserved before
it is rendered. 

Note: the HTML control already uses the same approach.

Also added a test case and cleaned up the AMD references.

Change-Id: I8516162aa3542f29a6ae80d1acbba2f01f4ee52f
BCP: 1580086102