[FIX] jQuery.sap.global: log nested err stack for module loading issues
For better traceability of module loading issues we concat the error
stacks for eval the module content and for the require call. This allows
to indentify the root cause of module loading issues in the log without
additional debugging.

For all browsers except of Chrome we log the error message including the
stacktrace which caused the issue in the console to see the source of
the issue properly.

Change-Id: I4728c13ab051379584419e0d2b3ac5dbac271986