[FIX] Component: removed the extends.component detection for manifest
Since the detection of base components is not working properly the
sap.ui5/extends/components property could be filled with a Component
having no manifest.json. This is not the expected bahavior and has to be
avoided. Therefore this detection mechanism is removed.

Change-Id: I5e9b6ed41ac9388f509848fcf42b5196668b8a50
(cherry picked from commit a163a7b61137f34b3209174e97cc4556070727ab)
CR-Id: 012003146900003341922015