[FIX] sap.m.Popover: correct placement when calcBest is used
When Popover size exceeds horizontal and vertical space around the opener
control, the calcBest position method is used. In this is performed a
check on which side there is the most available space to place the
Popover.

If left or right is used this placement position should be switched to the
oposite if you are in RTL mode. For example if you are in LTR mode and the
side with the most space is Right, the _oCalcedPos should be Right.
However, if the application is in RTL mode this Right, should be switched
to Left.

BCP: 1580096523
Change-Id: Ia30208523f7b5fe43a4a2dd25a3e17ed635e13a1