[FIX] sap.m.Popover: Clear old styles for the arrow positioning
Clear styles for positioning Popover's arrow. This is needed because if
the Popover uses the Auto placement and originally it is placed for
example on the Right and after calcBest position, the new placement is
Bottom, the arrow has inline styles for top and right, but it only needs
the right styles.

BCP: 1580096523
Change-Id: I53b011abfb652ad71595c2d8ad4046a4265ba7c6