[INTERNAL] sap.m.OverflowToolbar - optimization for SearchField
This change forces sap.m.SearchField to always have selectOnFocus=false when in the OverflowToolbar.
Thanks to this the cursor is always repositioned at the end of the SearchField when the toolbar is invalidated and the SearchField - refocused.
In the future it's possible to add a _change listener for SearchFields to help track the current state of selectOnFocus, because now the variable
_origSelectOnFocus is not synced with external changes to the property.

Change-Id: Iacc57486d2605e5e1af02ac134097449a4115193
CR-Id: 012003146900003658442015
(cherry picked from commit 20fb60de90132413b2c9c7e2b7368b30519219e7)