[INTERNAL] sap.m.OverflowToolbar - optimization for SearchField
This change forces sap.m.SearchField to always have selectOnFocus=false when in the OverflowToolbar.
Thanks to this the cursor is always repositioned at the end of the SearchField when the toolbar is invalidated and the SearchField - refocused.
In the future it's possible to add a _change listener for SearchFields to help track the current state of selectOnFocus, because now the variable
_origSelectOnFocus is not synced with external changes to the property.

Change-Id: I68930d5c35025ba03c31f5dd3d2d8c1668b2c3ac
CR-Id: 012003146900003658432015