[FIX] ODataMetadata: Catch Rejected Promise for Initial Metadata Load
The initial loading of the metadata document creates a promise which
is not referenced in the ODataMetadata object. In case this initial
loading fails, an "unaught reject" is logged to the console.
Since the error handling is done through eventing mechanisms, this
initial reject can be safely ignored. In case metadata loading is
started manually afterwards by menas of the refreshMetadata()-method
on the ODataModel, the returned promise rejects in case of an error
just as before.

BCP: 1580104700
CR-Id: 012003146900003838062015
Change-Id: Ida515ece78472cb96df3f3a59026fc3958a8678c