[INTERNAL] Japanese: Change signature of constructor and UTC method
For better compatibility with the other Date classes, the Japanese no
longer has the era as first parameter, but the year parameter is the
first parameter and will be treated as year in the current emperor era,
if less then 100 and gregorian year of greater than 100. To pass era and
year to the constructor, an array has to be given as year parameter.

Change-Id: I3a03d687d906cb2d5bb8a1cf131c1f1debf8ac06