[INTERNAL] jQuery.sap.global: XHR fix for jQuery issue #2068
With jQuery 1.11.1 the creation of the XHR object has been changed so
that it validates the supported HTTP methods and decides whether to
create a standard XHR or and ActiveXObject. In this function the methods
PATCH and MERGE are not listed and so the fallback to ActiveXObject
takes place.

Change-Id: Id46323b3774d351b8b9c5640d520e17b0e50d5dd
BCP: 1580101869
(cherry picked from commit ec7ff17e9bdad45f46f9df7f2a97838469197bed)
CR-Id: 012003146900004362572015