[INTERNAL] Add library / dist version to library.css URI
To enable custom theme compilation on demand, the library / dist
version needs to be passed to the server.

This change adds a new "sap-ui-versionedLibCss" configuration.
The core bootstrap loads the "sap-ui-version.json" if the
configuration is set to "true". This request will be sync/async
according to the "sap-ui-preload" configuration option
(sync by default).

The following theme resource will then be loaded with
the library and dist version as a URI parameter:
- library(-RTL).css (or any other variation)
- library-parameters.json

URLs within the CSS or Parameters are not modified.

CR-Id: 012003146900003980482015
(cherry picked from commit 26aa409e745fd73ef1983bd8149047a14be57c48)
Change-Id: I3c6a00117aa894c159df3330853cde14c2538d00