[INTERNAL] sap.m.OverflowToolbar - disable caching for dialog
This change introduces the possibility to disable caching of control sizes and does so for the dialog.
Furthermore, last-cached control widths are preserved between rerenderings so that elements with width "null",
i.e. the ones that were in the overflow, are counted with the width they had the last time they were in the toolbar.

Change-Id: Ib1f6c66ba9d924b34f44dbb48d7b001f9b024bd5