[FEATURE] Component: config options for manifest first loading behavior
This feature allows to enable the manifest first loading behavior for
Components. This means that instead of loading the Component Controller
directly the Component factory will now load the manifest.json and
preload the dependencies and libraries as well as registers the defined
resource roots of the Component. Before the only option to preload the
dependencies was the usage of the asyncHints options for the Component
factory.

The manifest first will work for all new Components with a manifest.json
which is available as a separate file. In all other cases there will be
a 404 penalty when the factory tries to load the manifest.json.
Therefore the manifest first loading behavior can be enabled in general
via the configuration option "manifestFirst" as part of the UI5 core
configuration and as well defined as parameter for the Component
factory.

Global configuration:
  <script id="sap-ui-bootstrap"...
    data-sap-ui-manifestFirst="true"></script>

Component factory:
  sap.ui.component({manifestFirst: true});

By default the manifest first loading is still disabled.

Change-Id: I5a61ba3aa29258d4b1186475fd1eecc60b7ee12c