[FIX] ODataMessageParser: Remove Double Errors
When the service returns an error-document there is no way in
OData v2 to return a target for the first error. In order to
set a target, the error message needs to be repeated in the 
errordetails array at the first position.
Until now the target of the two errors was
compared, which makes no sense since the first one can,
by definition, have no target on OData v2.
This change removes the target check.

BCP: 1580164127
Change-Id: I8aa5c23e13fa53d300b2dafd9135539756976acb