[FIX] sap.m.Toolbar - margins when last child is invisible fixed
Toolbar uses first-child/last-child selectors for items margins.
For this reason, when the first or last child is invisible, the design
looks broken.
The aim of this fix is to rewrite the css in such a way that first/last
child selectors are not needed.
They are still in the files in order to override the ones from
BarBase.less.

Change-Id: I7bac410dca80c8a3907444cbe9cb6fe46f5e3ecb