[FEATURE] jQuery.sap.FrameOptions: Allow service config via <meta> tag
This change will allow to configure a "whitelistService" via a meta element
placed in the HTML <head> before the UI5 bootstrap script:

<meta name="sap.whitelistService" content="/my/whitelistservice/url" />

It will only use the service URL if it wasn't configured via the standard UI5
configuration. Furthermore the "frameOptions" configuration will be set to
"trusted" if not explicitly defined.

Change-Id: I37f8305786346dbd963257982e789deeb40e76b9