ODataModelAdapter - allow injection of OData4SAPAnalyticalModel instance
New method setAnalyticalExtensions allows injection of an already
existing OData4SAPAnalyticalModel instance. Normally, the adapter
creates such an instance when it is applied to an OData model. However,
this means that the service metadata is parsed and transformed more than
once, which should be avoided, because it is a complex operation.

Change-Id: Ib7156d05ea8bf28ded9319260c9dac5de24437c8