[FIX] ES6-Promise Polyfill replaced native implementation in Safari
The ES6-Promise polyfill is meant to replace any non-compliant Promise
implementation upon loading the module.
We already changed the code before to accomodate Chrome, which has a
stange toString-conversion on its Promise implementation.
For Safari we had to remove the negative check for Promise.cast as
Safari now implements a cast method on the promise.

BCP: 1580157108
Change-Id: I0fc0f429bd31701d87b82aa9026823e22c2e3ca4