[INTERNAL][FIX] sap.ui.core.ws.WebSocket: QUnit test with https
The test wasn't working on a https server as the browser blocks
unsecure connections (ws) over secure sites (https).

As the socket endpoint doesn't matter for this test, it's the best
to just make it relative to the page, which will then use the same
protocol (secure/unsecure) depending on the http server.

Change-Id: Idb654e9a57d67242c89cfd152a74a6d507faa7ad
BCP: 1580219622