[FIX] ODataMessageParser: Specification changes
For NavigationProperties:
If a NavigationProperty is requested, the messages for the "parent"
Entity should not be deleted.
Example: GET /Products(1)/toSupplier 
 - used to delete all Messages for /Products(1)

For Entities:
The "parent" EntitySet should not be part of the affected targets,
meaning that messages for the entire collection should not be deleted
just because one entry of that selection has been requested.
Before this all messages for the parent collection were deleted when an
entry returned anything.
This only concerns messages for the EntitySet itself, not for its
entities. 
Example: GET /Products(1)
  - used to delete all messages for /Products(1) and /Products (but not
    /Products(2), Products(3), etc)
  - now it only deletes all messages for the single entity /Products(1)

Change-Id: I98ff099c73f8ca5326132f0c164918f26ce7cee5