[FIX] DataState: Repair DataState Propagation
Messages where initially set to null in the DataState object, which
lead to a DataState-change-event when they were set to []. DataState
now always returns an Array for messages, which can be tested by
accessing its length property

DataState Changes have been calculated using the last value per 
property, thus losing the difference to the initial state changing
value from "A" to "B" to "A" would report a change from "A" to "A".

Firing the DataStateChange lead to the DataState being flagged as
unchanged, which lead to erasing the differences on the next update of
a property, which could be at any given time in the future. Now the
changed properties are cleared after the binding has fired the
CombinedDataStateChange asynchronously.

Changes to sap.m.InputBase:

When a ValueState has been set by the application, it was overwritten
with every DataState change, regardless whether messages were
propagated or not. This change makes sure messages changed before
changing the ValueState.

BCP: 1580181600
BCP: 1580224764
BCP: 1580225792
BCP: 1570854104
BCP: 1580109193
BCP: 1570846075
BCP: 1570843888
BCP: 1570867530
Change-Id: I3fb265436380781a8270eb2fe0ca5a717148b341