[FIX] Locale: accept "DIGIT 3alphanum" variants, multiple variants
Language tags with a 4-character variant (like in de-CH-1901) have been
rejected before. And for multiple variants, the getter only returned the
last variant, same for extensions with different singletons.

The same fix also had to be applied in jquery.sap.resources as it 
contains a copy of the BCP47 regexp to avoid cross layer dependencies.

Also migrate affected QUnit tests to QUnit 2.0 APIs.

Change-Id: I506da3d49ddaf93c07dbacaf03c766a076174fe6