[INTERNAL] QUnit TestRunner modifications
1. Moved the "oContext" into the find() function to reset the list of found
tests from the testsuite. Until now, if the testsuite to lookup was
changed in the TestRunner UI, the result for found QUnit tests was just
extended, if there was a resultset already available.
With this change, the resultset is cleared for every testsuite leading
to display only the QUnit tests which are relevant for this testsuite.

2. Moved the closing of test selection area into the run() function.
This is necessary, as the solution before, only closed the test
selection area, when starting the tests manually by clikcing on the
"Run" button. If tests were executed automatically, e.g. by using the
autostart URL parameter, the test selection area remained open.

CR-Id: 012006153200004763812015
Change-Id: Iaf98e4b233f21e5d68ef38489825359476f548f8