[INTERNAL] sap.m.ComboBox: remove unnecessary flag in the keydown event
Do not unnecessary mark the keydown event as handled. The combo box is
handling the keydown event, but there is no need to mark the event as
handled, to let containers react to this event. e.g.:

When the combo box is render inside a table and the focus is in the
combo box input field and F7 is pressed, the table should react on this
event.


Change-Id: Ibe330b00be1493e650dd9fafa38124827a0f32d9
CR-Id: 012003146900004878262015