[INTERNAL] sap.ui.core: update coverage excludes (sap-ui-*.js)
The instrumentation logic has been changed to instrument all single
files and use them to re-bundle "sap-ui-*.js" files. This means that
the reporting will not show "sap-ui-*.js" files as source files,
so they need to be excluded here. Otherwise they would be listed
additionally to the actual files with a coverage of 0%.

Change-Id: I770808059c6ed2ebf411703bd247543edff0bcc7