Expression binding meets Control Binding
Add possibility to bind one controls property to another controls property.

For example it is now possible to bind the pressed state of a ToggleButton to the enabled property of some form controls.
Therefor I added a new prefix @ for bindings, to mark a binding as an control binding. I have chosen the @ sign because it is mostly used within the same view to refer at other controls.
The Binding Path is therefor produced like that: {=${@ControlId>Property}}.
For example <Input enabled='{=${@MyToggleButtonId>pressed}}'></Input>)

Currently id resolution is done by sap.ui.getCore().byId call and therefor only by setting an absolute id.
But if you like the feature and think about taking it over to one of your next releases, I can also try to implement a relative resolution.

Some tests to show the behavior:
http://localhost:8080/testsuite/test-resources/sap/ui/core/qunit/ExpressionParser.qunit.html?sap-ui-debug=true&sap-ui-language=de-DE&sap-ui-theme=sap_bluecrystal&sap-ui-accessibility=true&sap-ui-jqueryversion=1.11.1&module=Control%20Binding