[INTERNAL] Manifest: refactoring logic to detect component name
When loading a component with a component name this component name can
already be used to set the component name within the manifest. In case
of development scenarios this can override the component name defined in
sap.ui5/componentName or sap.app/id which could contain a placeholder
for resource filtering during build.

Change-Id: I7889924224f8828b64047f435fcb90f7a821f165
BCP: 1570848177
BCP: 1580183307
CR-Id: 012006153200004769182015