[FEATURE] UIComponent: added option in manifest to enable autoPrefixId
A UIComponent is able to enable the auto prefixing for IDs of
ManagedObjects which are created in the context of the createContent
function. With that change this option can be enabled with the setting
sap.ui5/autoPrefixId in the manifest.json.

This option is required in case of defining a View (sap.ui5/rootView)
and specifying an id for that View. In this scenario "not" using the
auto prefixing would lead to duplicate id issues when using the
Component multiple times.

Change-Id: Ic9ea38889621ebe6cf7598a8ae9a456aaa63aea8
(cherry picked from commit 6f40776bb11af83f58cce722ce5ba286463689fc)
CR-Id: 012006153200004742032015