[FEATURE] v2.ODataAnnotations: Optimized Annotations Loader
The new Annotations loader now has simpler API and event system.
All Annotation loading starts asynchronously after an URL has been
added as a new annotation source to the loader.
The Annotation parsing starts in the same order as the annotations
sources have been added.

The new ODataAnnotations class will be used to load the annotations
in the v2.ODataModel.

Change-Id: Iec96da6778346c11f940ca517f0fc36513a0c3de