[FIX] sap.m: add library info to semantic elements/controls
The library name needs to be provided for elements/controls
inside a sub-namespace as the automatic guessing assumes that
the containing library would be "sap.m.semantic" which is not
correct. This results to automatically registering it as a library
which could cause issues as the controls are not assigned to the
correct library.

Change-Id: Ida56179baf38a23d3f1d4a7ca699ac874c952524