[FIX] Theme Change: custom.css handling
This change fixes some issues related to the custom.css when
changing the theme during runtime.
The ThemeCheck did not check whether the stylesheet object assigned
to the <link> element is the new one.
This should now delay the custom.css handling until the theme has
been completely loaded.

Note: There is still an issue when loading the CSS from a different
origin as the rules can't be read and therefore the custom.css
handling doesn't happen.

CR-Id: 012006153200004916222015
BCP: 0020079747 0000897035 2015
Change-Id: Ie17a1741fa1cf03d3d255d50b879b1069409e588