[FEATURE] Customizing: enabled support for controller replacements
The controller replacement allow to specify another controller for a
given controller name. The controller will be replaced in the view
creation by checking the customizing configuration for such a controller
replacement.

The controller replacement can be specified in the ComponentMetadata in
the section /customizing/sap.ui.controllerReplacements or in the
manifest.json in the section
/sap.ui5/extends/extensions/sap.ui.controllerReplacements. The value of
the entry controllerReplacements is a map where the key is the name of
the old controller and the value is the name of the new controller which
should be loaded instead.

Change-Id: Ia7396699a9d3f4530a4b51ddc768735f9573df31