[INTERNAL] ComponentContainer: properly manage Component lifecycle
The ComponentContainer can now manage the lifecycle of a Component
which has been created by the ComponentContainer. Therefore a new
property "lifecycle" has been added which can be used to define that the
ComponentContainer should manage the lifecycle of the Component and
destroy it once another Component is set or the ComponentContainer is
destroyed. The default will be the "legacy" lifecycle which keeps the
previous behavior to destroy the associated Component once the
ComponentContainer is destroyed.

Change-Id: Ie1ea697db5e6422ff6e5c11d8132da0734a5e688