[INTERNAL] escapeRegExp: escape minimal set of characters
The previously escaped characters ',' '-' and '#' only have a meaning 
in regular expressions when they occur after some other special
character. As long as those 'other' characters are properly escaped, it
is not necessary to escape the above 3 characters, too. The same holds
btw. true for ':'.

Whitespace generally doesn't have to be escaped.

Change-Id: I5f140924c36fb2bbdea20187dc4272597c99cd03