[INTERNAL] Use api.json in explored app, improve formatting of JSDoc
The new api.json files are used as a third (and preferred) source of
information for the settings (properties, aggregations, associations,
events) and methods / fields of a class or control.

Only if the api.json cannot be determined or found, the old sources
(metamodel files or source code) are accessed.

Not all methods that could be displayed are indeed displayed (see
change in entity.controller.js). This limitation was added to give the
user a better overview about the 'core' methods. Displaying all get and
set methods for properties seemed to confuse more than that it helps.
Nevertheless, the explored app might be enhanced to show that
information on demand (same for static methods which are also 
suppressed)

Additionally, long missing formatting is applied to the displayed
floating text. The improvements include 

 - inline {@link something} tags now become real links (although not 
   all targets can be supported by the explored app, e.g. one can't 
   navigate to a specific method or event)
 - text is splitted into paragraphs as intended by the authors

 
Change-Id: I2a26028f7ac75c6f6b50e861aa3b5e2eb39aea14