[FEATURE] Upgrade QUnit to version 1.18.0
AMD registration of QUnit has been disabled.

In most of our test cases, QUnit is loaded via a script tag. If
requireJS has been loaded before, (e.g. after chart library has been
loaded), it rejects to define QUnit without a module name and throws a
runtime error.

Furthermore, preventing the autorun in that case also causes problems,
as tests are not prepared to handle this (they don't call QUnit.start).

To make the migration to QUnit 1.18 (which is necessary to benefit from
some fixes in QUnit) as smooth as possible, the AMD registration is 
commented out.

Drawback: the packaged QUnit module can't be loaded via requireJS or
another AMD loader.  Currently this isn't needed in typical
UI5 scenarios. As an alternative, tests could load their own copy of
QUnit or load it via script tag and redeclare it in QUnit afterwards or
configure a 'shim' for it in requireJS.

Change-Id: I305a5a164ec4ef68013774e08d79cf0a8123c762
CR-Id: 012006153200000001132016