[INTERNAL] Remove duplicate file whose name only differs in case
TimelineOverview.js was mistakenly added twice to the repo, once with 
uppercase 'L', and once with lowercase 'l'. On case-insensitive file
systems (like NTFS) this causes trouble as only one version of the file
can be synced and the other is regarded as deleted (at least by EGit).

This change removes the outdated version with the upper case 'L'. 
After syncing to this new change, it might be necessary to do a hard
reset to the HEAD or to do a checkout -f to enforce the proper spelling 
of the remaining file.

Change-Id: Ia231ab19df9db3ef03baa13cd9841226e2afa010