[FIX] Component/Manifest: Read manifest entries
The logic in the Manifest "getObject" method was a
bit strange regarding handling of falsy values.
It was not possible to access those as they have
been converted to "null".

With this change the method returns the values as
defined and prevents accessing native properties/methods.

CR-Id: 012006153200000333202016
Change-Id: I059f11868fb6c83ae5ac958808d5ece29f99195d
(cherry picked from commit 9cad2b2cf6af9af074a4010dee416636a54a25d2)