implement global window.onerror handler
A listener is chained to window.onerror at the start of core by
including the globalErrorHandler.js file in grunt/config/concat.js.

This functionality is opt-in using the xx-enableGlobalErrorHandler
configuration property introduced in sap.ui.core.Configuration.js.

sap.m.ErrorDialog is introduced to show error details in a textarea.

When an unchecked error occurs, an attempt to render sap.m.ErrorDialog
is made. If that fails, then a native JS dialog is rendered with the same
error information inside a textarea.

The native JS dialog is entirely self-contained within the core file so
no additional requests need to be made (styling, message bundles).