[FEATURE] v2.ODataModel: getObject with select and expand parameters
In the ODataModel we store entity data in an internal structure, where
all expanded entities are replaced with references. To allow the
application to access the data on the client in a more convenient way,
this change introduces the possibility to define expand and select
options, to chose which properties/navigation properties to include in
the data.

Change-Id: Ide414268aabe0ba7b39f4c154c355cb6706971b9
BCP: 1580241153
Fixes https://github.com/SAP/openui5/issues/720