[INTERNAL] QUnit: fix IE11 and Edge crash in special test cases
The QUnit test for jquery.sap.dom.js crashes on IE11/Edge when 
the option hidepassed is specified. This could be tracked down to 
a change in QUnit that makes hidden test results "position:absolute",
but without a parent element whose position is non-static.

Forcing the qunit-tests div to "position:relative" fixed the issue.

Change-Id: I2add0fad816777f7981a7a12608b6ad26ce65a3c
BCP: 1670034310