[INTERNAL] build: only create core files if lib is included
When running the grunt "build" task with the "--libs" option the
merged core files (e.g. "sap-ui-core.js") have been created even
when the "sap.ui.core" library was not included in the list of
libraries to be built.
This change fixes that issue and only creates the files when the
core library is part of the set of libraries.

Related issue: https://github.com/SAP/openui5/issues/757

Change-Id: I307a62ef9c8740b3d644f5addf2004ec37c45ed6