[FIX] sap.ui.commons.TriStateCheckBox: fix css img urls
Some img urls have been defined in the "base" theme although they
are only present in actual themes (sap_hcb, sap_bluecrystal,
sap_goldreflection).
This change moves the background-image definitions from base to
the actual theme so that the relative url points to the right path.

Furthermore the "TriStateCheckBox.png" image is referenced in the
"base" theme where also the image is located.
In the themes "sap_bluecrystal" and "sap_goldreflection" the exact
same image is located and can be removed as the CSS refers to the
image in the "base" theme.

Change-Id: I6c269508a7d10e72f481ad3b1ba9a7a9612e7c0a