[INTERNAL] sap.m.ComboBoxBase, sap.m.Select: always open the dropdown list
Eliminate unnecessary checks when opening the dropdown list.
Previously, the dropdown list was only opened if it was not empty.
However, this is an edge case, because in most of the cases,
the dropdown list has some items in it.

In this commit, we remove some unnecessary check to always open the
dropdown list when e.g.: the dropdown button is press to better support
lazy loading features.

Change-Id: I29c9d710f0852c7ddf8efc288d8b966e2305f54d