[FEATURE] Component: forward instance Manifest to Component in factory
The Component factory now returns a scoped function which keeps the
reference to the instance specific Manifest which can be delegated to
the Component instance once it is created. This enhancement is needed to
decouple the load of the Component class and the instantiation of the
Component when using the manifest-first mode.

Change-Id: I3b885c28913ecf76a4f3dbb790acec18366820d3