[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Removed redundant form spacing
Removed the extra left padding of form content ffor forms that appear within the blocks of the ObjectPageLayout. The change is requested from UX VD. The reason for the change is that the forms are already placed in a grid cell that provides the required horizontal spacing, so the extra spacing for form content is redundant.

CR-Id: 012006153200001016422016
Change-Id: Ic6678b52e75e93587f3c8896da6139a97395251c
(cherry picked from commit d4b254e6bf8512f17d8f7152529f9403fadc2d2c)