[INTERNAL][FEATURE] Component: early model creation (experimental)
This feature enables to create models declared in the manifest of
a component during preload of the component's dependencies which
should result into better parallel loading of files required to
run the component.
Especially when requesting an OData $metadata document this
should save some time during startup.

Currently it is only active when the experimental URI parameter
"sap-ui-xx-preload-component-models" is set to "true".

A model will be created directly after loading the component's
manifest.json in case
- model class is already loaded/declared
- model is not a sap.ui.model.resource.ResourceModel as it
  usually contains files which are part of the Component-preload
  which would then result into loading them twice.

All other models will be created on component init just like before.

It still needs to be discussed how this can be activated/deactivated
for specific models / model instances.

Change-Id: Id1ae7e6564e406af549fa0f066571ace4bec50fa