[FIX] v2.ODataModel: Call error handler in case of SAML redirect
In case the session is expired (e.g. after not interacting with an
application for some time) the next OData request will be redirected to
the identity provider for reauthentication. If this request is sent by
the ODataModel, it will receive a response with 200 and HTML content.
This must be handled gracefully by the ODataModel, so should result in
calling the error handler, instead of throwing a script error.

BCP: 1570873819
BCP: 1570881819
Change-Id: I61a96bc2bd6c0f1049c2cca35138ad85ab2f2080
CR-Id: 012003146900000069142016